partially completed problem generator
