+++ src/components/ProblemGenerator/QuestionGeneratorDisplay.jsx
@@ -3,51 +3,31 @@ import Timer from "./Timer";
 import Question from "./Question";
 
 const QuestionGeneratorDisplay = () => {
-  const [category, setCategory] = useState("fundamentalPeriod");
+  const categories = [
+    "Fundamental Period",
+    "Category 2",
+    "Category 3",
+    "Category 4",
+    "Category 5",
+    "Category 6",
+  ];
+  const [category, setCategory] = useState("Fundamental Period");
   return (
     <div className="w-2/3 m-auto">
       <div className="grid grid-cols-3 gap-5 m-5 ">
-        <div
-          className="border-2 border-black rounded cursor-pointer"
-          onClick={() => setCategory("fundamentalPeriod")}
-        >
-          Fundamental Period
-        </div>
-        <div
-          className="border-2 border-black rounded cursor-pointer"
-          onClick={(e) => setCategory(e.target.value)}
-        >
-          Category 2
-        </div>
-        <div
-          className="border-2 border-black rounded cursor-pointer"
-          onClick={(e) => setCategory(e.target.value)}
-        >
-          Category 3
-        </div>
-        <div
-          className="border-2 border-black rounded cursor-pointer"
-          onClick={(e) => setCategory(e.target.value)}
-        >
-          Category 4
-        </div>
-        <div
-          className="border-2 border-black rounded cursor-pointer"
-          onClick={(e) => setCategory(e.target.value)}
-        >
-          Category 5
-        </div>
-        <div
-          className="border-2 border-black rounded cursor-pointer"
-          onClick={(e) => setCategory(e.target.value)}
-        >
-          Category 6
-        </div>
+        {categories.map((item) => (
+          <div
+            className="border-2 rounded-lg cursor-pointer"
+            onClick={(e) => setCategory(e.target.value)}
+          >
+            {item}
+          </div>
+        ))}
       </div>
       <div className="flex justify-around">
-        <div className="border-2 border-black rounded-full w-40">Current</div>
+        <div className="border-2 rounded-full shadow-xl w-40">Current</div>
         <Timer />
-        <div className="border-2 border-black rounded-full w-40">Best</div>
+        <div className="border-2  shadow-xl rounded-full w-40">Best</div>
       </div>
       <Question category={category} />
     </div>
